Refactor Config classes into Template classes with revised hierarchy
Replaces the HintServiceMeow.UI.Models.Config namespace with
HintServiceMeow.UI.Models.Template and redesigns the class hierarchy:

Hierarchy changes:
- AbstractHintConfig removed; replaced by AbstractHintTemplate
  (SyncSpeed, FontSize, LineHeight, Text — no Id/Hide).
- HintConfig split into AnonymousHintTemplate (positional blueprint,
  no Id/Hide) and HintTemplate (extends anonymous with Id + Hide).
- DynamicHintConfig split into AnonymousDynamicHintTemplate (layout
  blueprint, no Id/Hide) and DynamicHintTemplate (extends anonymous
  with Id + Hide).
- HintPositionConfig -> HintPositionTemplate (standalone, unchanged).
- DynamicHintPositionConfig -> DynamicHintPositionTemplate (standalone,
  unchanged).

Method renames:
- ApplyTo(T) -> ApplyTemplate(T) in all classes.
- ToHint() -> GetHint() (inherited via virtual dispatch on ApplyTemplate).
- ToDynamicHint() -> GetDynamicHint() (same pattern).

Anonymous templates allow applying a shared visual/structural style to
multiple hints without affecting their Id or Hide state. Full templates
(HintTemplate, DynamicHintTemplate) inherit from anonymous counterparts
and override ApplyTemplate to additionally set Id and Hide.
