+++ HintServiceMeow/UI/Models/Template/DynamicHintPositionTemplate.cs
@@ -1,19 +1,23 @@
-namespace HintServiceMeow.UI.Models.Config
+namespace HintServiceMeow.UI.Models.Template
 {
-    using HintServiceMeow.Core.Enum;
     using HintServiceMeow.Core.Models.Hints;
 
     /// <summary>
-    /// Full configuration class for a <see cref="DynamicHint"/> (auto-positioning hint).
-    /// Inherits the common base properties from <see cref="AbstractHintConfig"/> and adds
-    /// all layout, boundary, margin, priority, and strategy properties specific to
-    /// <see cref="DynamicHint"/>.
+    /// A lightweight, layout-only template for a <see cref="DynamicHint"/>. Contains only
+    /// the boundary, target coordinate, and margin properties; carries no content, font,
+    /// priority, strategy, identity, or visibility state.
     /// <para>
-    /// Designed for serialization (JSON/YAML). All properties are nullable so that only
-    /// explicitly set values are applied when using <see cref="ApplyTo"/>.
+    /// Use this template when you only need to adjust the layout region of an existing
+    /// dynamic hint without touching any other settings.
+    /// Intentionally does <b>not</b> inherit from <see cref="AbstractHintTemplate"/> to
+    /// keep the surface area minimal.
+    /// </para>
+    /// <para>
+    /// All properties are nullable. Only non-null values are applied by
+    /// <see cref="ApplyTemplate"/>.
     /// </para>
     /// </summary>
-    public class DynamicHintConfig : AbstractHintConfig
+    public class DynamicHintPositionTemplate
     {
         // ── Boundaries ────────────────────────────────────────────────────────────
 
@@ -83,35 +87,16 @@ namespace HintServiceMeow.UI.Models.Config
         /// </summary>
         public float? RightMargin { get; set; }
 
-        // ── Behaviour ─────────────────────────────────────────────────────────────
-
-        /// <summary>
-        /// Gets or sets the display priority of the hint. Higher priority hints are less
-        /// likely to be displaced by other hints.
-        /// Maps to <see cref="DynamicHint.Priority"/>.
-        /// </summary>
-        public HintPriority? Priority { get; set; }
-
-        /// <summary>
-        /// Gets or sets the fallback strategy applied when no valid display position is
-        /// available within the hint's boundaries.
-        /// Maps to <see cref="DynamicHint.Strategy"/>.
-        /// </summary>
-        public DynamicHintStrategy? Strategy { get; set; }
-
         // ── Methods ───────────────────────────────────────────────────────────────
 
         /// <summary>
-        /// Applies the non-null properties of this config to the given <paramref name="hint"/>.
-        /// Only properties that have been explicitly set (non-null) are written; existing
-        /// hint values are preserved for any property that remains null.
+        /// Applies the non-null layout properties of this template to the given
+        /// <paramref name="hint"/>. Existing hint values are preserved for any property
+        /// that remains null.
         /// </summary>
-        /// <param name="hint">The target <see cref="DynamicHint"/> to update.</param>
-        public void ApplyTo(DynamicHint hint)
+        /// <param name="hint">The target <see cref="DynamicHint"/> to reposition.</param>
+        public void ApplyTemplate(DynamicHint hint)
         {
-            // Apply common base properties (Id, SyncSpeed, FontSize, LineHeight, Text, Hide).
-            ApplyBaseTo(hint);
-
             if (TopBoundary.HasValue)
                 hint.TopBoundary = TopBoundary.Value;
 
@@ -141,25 +126,6 @@ namespace HintServiceMeow.UI.Models.Config
 
             if (RightMargin.HasValue)
                 hint.RightMargin = RightMargin.Value;
-
-            if (Priority.HasValue)
-                hint.Priority = Priority.Value;
-
-            if (Strategy.HasValue)
-                hint.Strategy = Strategy.Value;
-        }
-
-        /// <summary>
-        /// Creates a new <see cref="DynamicHint"/> instance and populates it with all non-null
-        /// properties from this config. Properties that are null fall back to the hint's
-        /// built-in default values.
-        /// </summary>
-        /// <returns>A new <see cref="DynamicHint"/> configured from this instance.</returns>
-        public DynamicHint ToDynamicHint()
-        {
-            var hint = new DynamicHint();
-            ApplyTo(hint);
-            return hint;
         }
     }
 }
